Import of CJFIELDS/BioPerl-1.6.0_1 from CPAN.
gitpan-cpan-distribution: BioPerl
gitpan-cpan-version:      1.6.0_1
gitpan-cpan-path:         CJFIELDS/BioPerl-1.6.0_1.tar.gz
gitpan-cpan-author:       CJFIELDS
gitpan-cpan-maturity:     developer

+++ Bio/Seq/LargePrimarySeq.pm
@@ -1,6 +1,8 @@
-# $Id: LargePrimarySeq.pm 10525 2006-09-26 22:03:22Z sendu $
+# $Id: LargePrimarySeq.pm 16123 2009-09-17 12:57:27Z cjfields $
 #
 # BioPerl module for Bio::Seq::LargePrimarySeq
+#
+# Please direct questions and support issues to <bioperl-l@bioperl.org> 
 #
 # Cared for by Ewan Birney <birney@ebi.ac.uk>
 #
@@ -45,6 +47,17 @@ of the Bioperl mailing lists.  Your participation is much appreciated.
   bioperl-l@bioperl.org                  - General discussion
   http://bioperl.org/wiki/Mailing_lists  - About the mailing lists
 
+=head2 Support 
+
+Please direct usage questions or support issues to the mailing list:
+
+I<bioperl-l@bioperl.org>
+
+rather than to the module maintainer directly. Many experienced and 
+reponsive experts will be able look at the problem and quickly 
+address it. Please include a thorough description of the problem 
+with code and data examples if at all possible.
+
 =head2 Reporting Bugs
 
 Report bugs to the Bioperl bug tracking system to help us keep track
@@ -84,14 +97,14 @@ sub new {
 
     my $seq = $params{'-seq'} || $params{'-SEQ'};
     if($seq ) {
-	delete $params{'-seq'};
-	delete $params{'-SEQ'};
+    delete $params{'-seq'};
+    delete $params{'-SEQ'};
     }
     my $self = $class->SUPER::new(%params);
     $self->_initialize_io(%params);
     my $tempdir = $self->tempdir( CLEANUP => 1);
     my ($tfh,$file) = $self->tempfile( DIR => $tempdir );
-
+    $self->{tempdir} = $tempdir;
     $tfh     && $self->_fh($tfh);
     $file    && $self->_filename($file);    
     $self->length(0);
@@ -137,9 +150,9 @@ sub seq {
    my ($self, $data) = @_;   
    if( defined $data ) {
        if( $self->length() == 0) {
-	   $self->add_sequence_as_string($data);
+       $self->add_sequence_as_string($data);
        } else { 
-	   $self->warn("Trying to reset the seq string, cannot do this with a LargePrimarySeq - must allocate a new object");
+       $self->warn("Trying to reset the seq string, cannot do this with a LargePrimarySeq - must allocate a new object");
        }
    } 
    return $self->subseq(1,$self->length);
@@ -165,48 +178,48 @@ sub subseq{
    if( ref($start) && $start->isa('Bio::LocationI') ) {
        my $loc = $start;
        if( $loc->length == 0 ) { 
-	   $self->warn("Expect location lengths to be > 0");
-	   return '';
+       $self->warn("Expect location lengths to be > 0");
+       return '';
        } elsif( $loc->end < $loc->start ) { 
-	   # what about circular seqs
-	   $self->warn("Expect location start to come before location end");
+       # what about circular seqs
+       $self->warn("Expect location start to come before location end");
        }
        my $seq = '';
        if( $loc->isa('Bio::Location::SplitLocationI') ) {
-	   foreach my $subloc ( $loc->sub_Location ) {
-	       if(! seek($fh,$subloc->start() - 1,0)) {
-		   $self->throw("Unable to seek on file $start:$end $!");
-	       }
-	       my $ret = read($fh, $string, $subloc->length());
-	       if( !defined $ret ) {
-		   $self->throw("Unable to read $start:$end $!");
-	       }
-	       if( $subloc->strand < 0 ) { 
-		   $string = Bio::PrimarySeq->new(-seq => $string)->revcom()->seq();
-	       }
-	       $seq .= $string;
-	   }
+       foreach my $subloc ( $loc->sub_Location ) {
+           if(! seek($fh,$subloc->start() - 1,0)) {
+           $self->throw("Unable to seek on file $start:$end $!");
+           }
+           my $ret = read($fh, $string, $subloc->length());
+           if( !defined $ret ) {
+           $self->throw("Unable to read $start:$end $!");
+           }
+           if( $subloc->strand < 0 ) { 
+           $string = Bio::PrimarySeq->new(-seq => $string)->revcom()->seq();
+           }
+           $seq .= $string;
+       }
        } else { 
-	   if(! seek($fh,$loc->start()-1,0)) {
-	       $self->throw("Unable to seek on file ".$loc->start.":".
-			    $loc->end ." $!");
-	   }
-	   my $ret = read($fh, $string, $loc->length());
-	   if( !defined $ret ) {
-	       $self->throw("Unable to read ".$loc->start.":".
-			    $loc->end ." $!");
-	   }
-	   $seq = $string;
+       if(! seek($fh,$loc->start()-1,0)) {
+           $self->throw("Unable to seek on file ".$loc->start.":".
+                $loc->end ." $!");
+       }
+       my $ret = read($fh, $string, $loc->length());
+       if( !defined $ret ) {
+           $self->throw("Unable to read ".$loc->start.":".
+                $loc->end ." $!");
+       }
+       $seq = $string;
        }
        if( defined $loc->strand && 
-	   $loc->strand < 0 ) { 
-	   $seq = Bio::PrimarySeq->new(-seq => $seq)->revcom()->seq();
+       $loc->strand < 0 ) { 
+       $seq = Bio::PrimarySeq->new(-seq => $seq)->revcom()->seq();
        }
        return $seq;
    }
    if( $start <= 0 || $end > $self->length ) {
        $self->throw("Attempting to get a subseq out of range $start:$end vs ".
-		    $self->length);
+            $self->length);
    }
    if( $end < $start ) {
        $self->throw("Attempting to subseq with end ($end) less than start ($start). To revcom use the revcom function with trunc");
@@ -294,8 +307,9 @@ sub DESTROY {
     my $fh = $self->_fh();
     close($fh) if( defined $fh );
     # this should be handled by Tempfile removal, but we'll unlink anyways.
-    unlink $self->_filename()
-        if defined $self->_filename() && -e $self->_filename;
+    unlink $self->_filename() if defined $self->_filename() && -e $self->_filename;
+	# remove tempdirs as well
+    rmdir $self->{tempdir} if defined $self->{tempdir} && -e $self->{tempdir};
     $self->SUPER::DESTROY();
 }
